+++ check/Data.List.check.hs
@@ -1,0 +1,3 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+module Test where
+import Data.List.Compat